fix: resolve dropped socket/HTTP responses and price-drop sweep crash
- socket service was missing the keepAliveTimeout/headersTimeout override
  already applied to the backend, causing Railway's edge proxy to reuse
  connections the socket service had silently closed (Node's 5s default
  keep-alive). This surfaced as constant "transport close" churn and
  dropped real-time events even though the triggering action succeeded.
- price-drop favorites sweep was casting a float threshold into an
  integer-inferred parameter, failing every run since deploy.
- DELETE /notifications/push-token was shadowed by the earlier
  DELETE /:notificationId route, so logout's token unregister always
  500'd with a UUID cast error.

+++ backend/workers/engagementAlerts.js
@@ -45,7 +45,7 @@ async function sweepFavoriteAlerts() {
       FROM favorites f
       JOIN products p ON p.id = f.product_id
       WHERE p.is_active = TRUE AND p.deleted_at IS NULL
-        AND p.price <= f.alert_price * (1 - $1)
+        AND p.price <= f.alert_price * (1 - $1::numeric)
         AND (f.price_notified_at IS NULL OR f.price_notified_at < NOW() - INTERVAL '${PRICE_NOTIFY_COOLDOWN_HOURS} hours')
       LIMIT 500
     `, [PRICE_DROP_THRESHOLD]);
